LLM agent assistant refactor
