+++ galapassistant/apps/assistant/services/generation.py
@@ -1,0 +1,55 @@
+import os
+from dotenv import load_dotenv
+from smolagents import HfApiModel, ToolCallingAgent
+
+from galapassistant.apps.assistant.services.retrieval import RetrieverTool
+
+
+load_dotenv()
+
+LLM_MODEL_NAME = "HuggingFaceH4/zephyr-7b-beta"
+
+class GenerationService:
+    """
+    Service class for generating responses using an LLM.
+    """
+    def __init__(self):
+        """
+        Initializes the GenerationService by creating an instance of the LLM.
+        """
+        try:
+            self.llm = HfApiModel(
+                model_id=LLM_MODEL_NAME,
+                token=os.getenv("HUGGINGFACEHUB_API_TOKEN"),
+                temperature=0.3
+            )
+        except Exception as e:
+            print(f"Error initializing LLM: {e}")
+            self.llm = None
+
+    def generate_answer(self, user_query: str) -> str:
+        """
+        Generate a response from the LLM based on the provided query.
+
+        Args:
+            query (str): The user's query.
+
+        Returns:
+            str: The response generated by the LLM, or an error message if something fails.
+        """
+        try:
+            if self.llm is None:
+                return "LLM is not properly initialized."
+
+            retriever = RetrieverTool()
+            agent = ToolCallingAgent(
+                tools=[retriever],
+                model=self.llm,
+                verbosity_level=2
+            )
+            response = agent.run(user_query)
+            return response
+        except Exception as e:
+            error_msg = f"Error generating answer: {e}"
+            print(error_msg)
+            return error_msg
