--- galapassistant/apps/assistant/services/llm_service.py
@@ -1,51 +1,0 @@
-import os
-from dotenv import load_dotenv
-from langchain_core.output_parsers import StrOutputParser
-from smolagents import HfApiModel, ToolCallingAgent
-
-from galapassistant.apps.assistant.services.retriever_service import RetrieverTool
-
-
-load_dotenv()
-
-LLM_MODEL_NAME = "HuggingFaceH4/zephyr-7b-beta" ##Other models: "MaziyarPanahi/calme-3.1-instruct-78b", "meta-llama/Llama-3.1-70B-Instruct"
-FILE_PATH = os.path.join(
-    os.path.dirname(__file__),
-    "..",
-    ".knowledge_base",
-    "the Origin of Species.txt"
-)
-
-class AssistantLLMService:
-    """
-    Service class for generating responses using an LLM.
-    """
-    def __init__(self):
-        """
-        Initializes the AssistantLLMService by creating an instance of the LLM.
-        """
-        self.llm = HfApiModel(
-            model_id=LLM_MODEL_NAME,
-            token=os.getenv("HUGGINGFACEHUB_API_TOKEN"),
-            temperature=0.3
-        )
-
-    def generate_answer(self, user_query: str) -> str:
-        """
-        Generate a response from the LLM based on the provided query.
-
-        Args:
-            query (str): The user's query.
-
-        Returns:
-            str: The response generated by the LLM.
-        """
-        retriever = RetrieverTool()
-        agent = ToolCallingAgent(
-            tools=[retriever],
-            model=self.llm,
-            verbosity_level=2
-        )
-        response = agent.run(user_query)
-
-        return response
